requirements.txt
ccxt>=2.0.0
pandas>=1.3.0
numpy>=1.24.3
matplotlib>=3.4.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio>=3.4.3
inquirer>=3.4.0
TA-Lib>=0.4.26
argparse>=1.4.0
PyQt5>=5.15.4
Flask==
pandas==
pandas_ta==
krakenex==
python-dotenv==
schedule==
openai==
ta==
arrow==
pykrakenapi== 
pydantic==